+++ zephyr-mcumgr-cli/src/main.rs
@@ -11,7 +11,6 @@ mod formatting;
 use std::time::Duration;
 
 use clap::Parser;
-use rand::distr::SampleString;
 use zephyr_mcumgr::{MCUmgrClient, client::UsbSerialError};
 
 use crate::errors::CliError;
@@ -59,13 +58,8 @@ fn cli_main() -> Result<(), CliError> {
     if let Some(group) = args.group {
         groups::run(&client, args.common, group)?;
     } else {
-        let random_message = rand::distr::Alphanumeric.sample_string(&mut rand::rng(), 16);
-        let response = client.os_echo(&random_message)?;
-        if random_message == response {
-            println!("Device alive and responsive.");
-        } else {
-            return Err(CliError::EchoFailed);
-        }
+        client.check_connection()?;
+        println!("Device alive and responsive.");
     }
 
     Ok(())